Introduce sham toiletdensity
There is a circular dependency between the generation steps for toiletdensity and toiletcollections. As a stopgap, I simply introduced a sham toiletdensity table. This allows us to create the sham, create the toiletcollections table (with the wrong densities), and then create the real density information (which hopefully is not corrupted by the missing density data from toiletcollections), and then re-create the toiletcollections table with the actual densities.

This is a stop-gap. It is a hack. It enables reproducible ETL in a manner that I believe the tables were originally generated.